Changed the type of all mask values to uint64

+++ h/kernel/BitMasks.h
@@ -6,25 +6,25 @@
 #define PROJECT_BITMASKS_H
 #include "../../lib/hw.h"
 
-static constexpr size_t maskBit(uint8 bit) {
-    return (size_t) 0x01 << (size_t) bit;
+static constexpr uint64 maskBit(uint8 bit) {
+    return (uint64) 0x01 << (uint64) bit;
 }
 
 namespace kernel {
     namespace BitMasks {
-        enum class sstatus : size_t {
+        enum class sstatus : uint64 {
             SPP = maskBit(8),
             SPIE = maskBit(5),
             SIE = maskBit(1),
         };
-        enum class scause : size_t{
+        enum class scause : uint64{
             BNT = maskBit(63)
         };
-        enum class sip : size_t {
+        enum class sip : uint64 {
             SEIP = maskBit(9),
             SSIP = maskBit(1),
         };
-        enum class sie : size_t {
+        enum class sie : uint64 {
             SEIE = maskBit(9),
             SSIE = maskBit(1),
         };
